more complicated, but more correct
